requirements.txt
python-telegram-bot==13.0
gspread==3.6.0